[pf-437] Export filter for challenges and cat algorithm logic into differents files
