+++ api/tests/integration/domain/strategies/SmartRandom_test.js
@@ -10,7 +10,7 @@ const KNOWLEDGE_ELEMENT_STATUS = {
   INVALIDATED: 'invalidated'
 };
 
-describe('Unit | Domain | Models | SmartRandom', () => {
+describe('Integration | Domain | Stategies | SmartRandom', () => {
 
   describe('#constructor', () => {
     it('should create a course with tubes', () => {
@@ -611,79 +611,5 @@ describe('Unit | Domain | Models | SmartRandom', () => {
       // then
       expect(nextChallenge).to.equal(null);
     });
-  });
-
-  describe('SmartRandom._filteredChallenges()', function() {
-    it('should not ask a question that targets a skill already assessed', function() {
-      // given
-      const [skill1, skill2, skill3] = domainBuilder.buildSkillCollection();
-
-      const targetProfile = new TargetProfile({ skills: [skill1, skill2, skill3] });
-
-      const challengeAssessingSkill1 = domainBuilder.buildChallenge({ skills: [skill1] });
-      const challengeAssessingSkill2 = domainBuilder.buildChallenge({ skills: [skill2] });
-      const anotherChallengeAssessingSkill2 = domainBuilder.buildChallenge({ skills: [skill2] });
-      const challengeAssessingSkill3 = domainBuilder.buildChallenge({ skills: [skill3] });
-
-      const answerCh1 = domainBuilder.buildAnswer({ challengeId: challengeAssessingSkill1.id, result: AnswerStatus.OK });
-      const answerCh2 = domainBuilder.buildAnswer({ challengeId: challengeAssessingSkill2.id, result: AnswerStatus.OK });
-      const answers = [answerCh1, answerCh2];
-      const knowledgeElements = [
-        domainBuilder.buildSmartPlacementKnowledgeElement({ skillId: skill1.id, status: KNOWLEDGE_ELEMENT_STATUS.VALIDATED, source: 'direct' }),
-        domainBuilder.buildSmartPlacementKnowledgeElement({ skillId: skill2.id, status: KNOWLEDGE_ELEMENT_STATUS.VALIDATED, source: 'direct' }),
-      ];
-
-      const challenges = [
-        challengeAssessingSkill1,
-        challengeAssessingSkill2,
-        anotherChallengeAssessingSkill2,
-        challengeAssessingSkill3,
-      ];
-
-      // when
-      const smartRandom = new SmartRandom({ answers, challenges, targetProfile, knowledgeElements });
-      const result = SmartRandom._filteredChallenges({
-        challenges: smartRandom.challenges,
-        knowledgeElements,
-        predictedLevel:smartRandom.predictedLevel,
-        targetProfile: smartRandom.targetProfile
-      });
-
-      // then
-      expect(result).to.deep.equal([challengeAssessingSkill3]);
-    });
-
-    context('when the selected challenges cover more skills than the defined target profile', () => {
-      it('should ignore the already answered challenges, even if they have non evaluated skills', function() {
-        // given
-        const [skill1, skill2] = domainBuilder.buildSkillCollection();
-
-        const targetProfile = domainBuilder.buildTargetProfile({ skills: [skill1] });
-
-        const challengeAssessingSkill1 = domainBuilder.buildChallenge({ skills: [skill1, skill2] });
-
-        const answerCh1 = domainBuilder.buildAnswer({ challengeId: challengeAssessingSkill1.id, result: AnswerStatus.OK });
-        const answers = [answerCh1];
-        const knowledgeElements = [
-          domainBuilder.buildSmartPlacementKnowledgeElement({ skillId: skill1.id, status: KNOWLEDGE_ELEMENT_STATUS.VALIDATED, source: 'direct' }),
-        ];
-        const challenges = [
-          challengeAssessingSkill1,
-        ];
-
-        // when
-        const smartRandom = new SmartRandom({ answers, challenges, targetProfile, knowledgeElements });
-        const result = SmartRandom._filteredChallenges({
-          challenges: smartRandom.challenges,
-          knowledgeElements,
-          predictedLevel:smartRandom.predictedLevel,
-          targetProfile: smartRandom.targetProfile,
-        });
-
-        // then
-        expect(result).to.deep.equal([]);
-      });
-    });
-
   });
 });
